on controller: add existing station validation, refactoring create route validation; on service: deleted redundant methods

+++ src/main/java/schedule/controller/StationController.java
@@ -7,7 +7,9 @@ import org.springframework.web.bind.annotation.PathVariable;
 import org.springframework.web.bind.annotation.RequestMapping;
 import org.springframework.web.bind.annotation.RequestMethod;
 import org.springframework.web.servlet.ModelAndView;
+import schedule.model.Schedule;
 import schedule.model.Station;
+import schedule.service.api.ScheduleService;
 import schedule.service.api.StationService;
 
 import java.util.List;
@@ -18,6 +20,7 @@ import java.util.Locale;
 public class StationController extends BaseController {
 
     private StationService stationService;
+    private ScheduleService scheduleService;
 
     @RequestMapping(value = "/list")
     public ModelAndView listOfStations(Locale locale) {
@@ -29,20 +32,6 @@ public class StationController extends BaseController {
 
         return modelAndView;
     }
-
-/*    @RequestMapping(value = "/add", method = RequestMethod.POST)
-    public ModelAndView addStation(@ModelAttribute Station station,
-                                   Locale locale) {
-        stationService.addStation(station);
-
-        List<Station> stations = stationService.getStations();
-
-        ModelAndView modelAndView = new ModelAndView("stations-list");
-        modelAndView.addObject("stations", stations);
-        modelAndView.addObject("message", getMessage("message.stations.create.success", locale));
-
-        return modelAndView;
-    }*/
 
     @RequestMapping(value = "/edit/{station.id}", method = RequestMethod.GET)
     public ModelAndView editStation(@PathVariable(value = "station.id") int stationId,
@@ -71,13 +60,23 @@ public class StationController extends BaseController {
     @RequestMapping(value = "/delete/{station.id}")
     public ModelAndView deleteStation(@PathVariable(value = "station.id") int stationId,
                                       Locale locale) {
-        stationService.deleteStation(stationId);
-
-        List<Station> stations = stationService.getStations();
-
         ModelAndView modelAndView = new ModelAndView("stations-list");
+        List<Station> stations;
+        List<Schedule> resultList;
+
+        resultList = scheduleService.findScheduleByStation(stationService.getStation(stationId));
+        if (resultList.isEmpty()) {
+            stationService.deleteStation(stationId);
+            modelAndView.addObject("message", getMessage("message.stations.delete.success", locale));
+            stations = stationService.getStations();
+            modelAndView.addObject("stations", stations);
+
+            return modelAndView;
+        }
+
+        stations = stationService.getStations();
+        modelAndView.addObject("message", getMessage("message.stations.delete.fail", locale));
         modelAndView.addObject("stations", stations);
-        modelAndView.addObject("message", getMessage("message.stations.delete.success", locale));
 
         return modelAndView;
     }
@@ -85,5 +84,10 @@ public class StationController extends BaseController {
     @Autowired
     public void setStationService(StationService stationService) {
         this.stationService = stationService;
+    }
+
+    @Autowired
+    public void setScheduleService(ScheduleService scheduleService) {
+        this.scheduleService = scheduleService;
     }
 }
